requirements.txt
aiofiles==24.1.0
aiohttpx==0.0.12
annotated-types==0.7.0
anyio==4.6.0
asttokens==2.4.1
async-lru==2.0.4
async_openai==0.0.52
asyncio==3.4.3
backoff==2.2.1
certifi==2024.8.30
charset-normalizer==3.3.2
comm==0.2.2
decorator==5.1.1
distro==1.9.0
executing==2.1.0
frozendict==2.4.5
gitdb==4.0.11
GitPython==3.1.41
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
ipython==8.28.0
ipywidgets==8.1.5
jedi==0.19.1
Jinja2==3.1.4
jiter==0.6.1
jupyterlab_widgets==3.0.13
lazyops==0.2.84
loguru==0.7.2
MarkupSafe==3.0.1
matplotlib-inline==0.1.7
ollama==0.3.3
openai==1.51.2
parso==0.8.4
pexpect==4.9.0
prompt_toolkit==3.0.48
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
ruff==0.6.9
setuptools==69.0.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
stack-data==0.6.3
tiktoken==0.8.0
tqdm==4.66.5
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.2.3
wcwidth==0.2.13
widgetsnbextension==4.0.13
